Upload scaling methodology and examples
In the archives (content identical), there are my attempts to scale this test to linear algebra problems. You can clearly see that GPT 5 rewards LLaMa 4 for solving common problems and punishes it for misunderstanding rare ones. You can also see that, when the distribution is Zipfian, the performance differences are most obvious. There is also my conversation with GPT 5 High and DeepSeek V3.2 from LMArena where we discussed how to solve both problems to scale this test design.